[refactor] 클래스 역할에 맞게 패키지 분리
