+++ src/main/java/lotto/Main.java
@@ -1,5 +1,8 @@
 package lotto;
 
+import lotto.controller.LottoMachine;
+import lotto.view.LottoView;
+
 public class Main {
 
     public static void main(String[] args) {
